requirement.txt
asgiref==3.2.5
asn1crypto==1.3.0
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
cryptography==2.8
Django==3.0.3
django-cors-headers==3.2.1
django-filter==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
idna==2.9
Markdown==3.2.1
mysqlclient==1.3.14
pycparser==2.20
PyMySQL==0.9.3
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
wincertstore==0.2
